+++ coders/plasma.c
@@ -94,9 +94,9 @@ static void PlasmaPixel(Image *image,double x,double y)
   q=GetImagePixels(image,(long) (x+0.5),(long) (y+0.5),1,1);
   if (q == (PixelPacket *) NULL)
     return;
-  q->red=(Quantum) (MaxRGB*rand()/RAND_MAX);
-  q->green=(Quantum) (MaxRGB*rand()/RAND_MAX);
-  q->blue=(Quantum) (MaxRGB*rand()/RAND_MAX);
+  q->red=(Quantum) ((double) MaxRGB*rand()/RAND_MAX);
+  q->green=(Quantum) ((double) MaxRGB*rand()/RAND_MAX);
+  q->blue=(Quantum) ((double) MaxRGB*rand()/RAND_MAX);
   (void) SyncImagePixels(image);
 }
