+++ magick/decorate.c
@@ -208,30 +208,30 @@ MagickExport Image *FrameImage(const Image *image,const FrameInfo *frame_info,
     Initialize 3D effects color.
   */
   matte=image->matte_color;
-  accentuate.red=(Quantum) (((MaxRGB-AccentuateModulate)*matte.red+
+  accentuate.red=(Quantum) ((double) ((MaxRGB-AccentuateModulate)*matte.red+
     (MaxRGB*AccentuateModulate))/MaxRGB+0.5);
-  accentuate.green=(Quantum) (((MaxRGB-AccentuateModulate)*matte.green+
+  accentuate.green=(Quantum) ((double) ((MaxRGB-AccentuateModulate)*matte.green+
     (MaxRGB*AccentuateModulate))/MaxRGB+0.5);
-  accentuate.blue=(Quantum) (((MaxRGB-AccentuateModulate)*matte.blue+
+  accentuate.blue=(Quantum) ((double) ((MaxRGB-AccentuateModulate)*matte.blue+
     (MaxRGB*AccentuateModulate))/MaxRGB+0.5);
-  accentuate.opacity=(Quantum) (((MaxRGB-AccentuateModulate)*matte.opacity+
-    (MaxRGB*AccentuateModulate))/MaxRGB+0.5);
-  highlight.red=(Quantum) (((MaxRGB-HighlightModulate)*matte.red+
+  accentuate.opacity=(Quantum) ((double) ((MaxRGB-AccentuateModulate)*
+    matte.opacity+(MaxRGB*AccentuateModulate))/MaxRGB+0.5);
+  highlight.red=(Quantum) ((double) ((MaxRGB-HighlightModulate)*matte.red+
     (MaxRGB*HighlightModulate))/MaxRGB+0.5);
-  highlight.green=(Quantum) (((MaxRGB-HighlightModulate)*matte.green+
+  highlight.green=(Quantum) ((double) ((MaxRGB-HighlightModulate)*matte.green+
     (MaxRGB*HighlightModulate))/MaxRGB+0.5);
-  highlight.blue=(Quantum) (((MaxRGB-HighlightModulate)*matte.blue+
+  highlight.blue=(Quantum) ((double) ((MaxRGB-HighlightModulate)*matte.blue+
     (MaxRGB*HighlightModulate))/MaxRGB+0.5);
-  highlight.opacity=(Quantum) (((MaxRGB-HighlightModulate)*matte.opacity+
-    (MaxRGB*HighlightModulate))/MaxRGB+0.5);
-  shadow.red=(Quantum) ((matte.red*ShadowModulate)/MaxRGB+0.5);
-  shadow.green=(Quantum) ((matte.green*ShadowModulate)/MaxRGB+0.5);
-  shadow.blue=(Quantum) ((matte.blue*ShadowModulate)/MaxRGB+0.5);
-  shadow.opacity=(Quantum) ((matte.opacity*ShadowModulate)/MaxRGB+0.5);
-  trough.red=(Quantum) ((matte.red*TroughModulate)/MaxRGB+0.5);
-  trough.green=(Quantum) ((matte.green*TroughModulate)/MaxRGB+0.5);
-  trough.blue=(Quantum) ((matte.blue*TroughModulate)/MaxRGB+0.5);
-  trough.opacity=(Quantum) ((matte.opacity*TroughModulate)/MaxRGB+0.5);
+  highlight.opacity=(Quantum) ((double) ((MaxRGB-HighlightModulate)*
+    matte.opacity+(MaxRGB*HighlightModulate))/MaxRGB+0.5);
+  shadow.red=(Quantum) ((double) (matte.red*ShadowModulate)/MaxRGB+0.5);
+  shadow.green=(Quantum) ((double) (matte.green*ShadowModulate)/MaxRGB+0.5);
+  shadow.blue=(Quantum) ((double) (matte.blue*ShadowModulate)/MaxRGB+0.5);
+  shadow.opacity=(Quantum) ((double) (matte.opacity*ShadowModulate)/MaxRGB+0.5);
+  trough.red=(Quantum) ((double) (matte.red*TroughModulate)/MaxRGB+0.5);
+  trough.green=(Quantum) ((double) (matte.green*TroughModulate)/MaxRGB+0.5);
+  trough.blue=(Quantum) ((double) (matte.blue*TroughModulate)/MaxRGB+0.5);
+  trough.opacity=(Quantum) ((double) (matte.opacity*TroughModulate)/MaxRGB+0.5);
   /*
     Draw top of ornamental border.
   */
